move utility funcs
